Moved `database.js` from `modal` folder
